test: consolidate n-api test addons - part2
It takes time to build each of the addons used to test n-api.
Consolidate a few of the smaller ones to save build time.

Get rid of one more small addon.

PR-URL: https://github.com/nodejs/node/pull/13380
