--- test/addons-napi/test_napi_status/binding.gyp
@@ -1,8 +1,0 @@
-{
-  "targets": [
-    {
-      "target_name": "test_napi_status",
-      "sources": [ "test_napi_status.cc" ]
-    }
-  ]
-}
